Include disjoint sets

+++ MeshFragments/Assets/Shaders/Compute/Fracturer/erodeGrid-comp.glsl
@@ -27,7 +27,7 @@ void main()
 	const uint index = gl_GlobalInvocationID.x;
 	if (index >= numCells) return;
 
-	bool isBoundary = bool((grid[index].value >> 7) & uint8_t(1));
+	bool isBoundary = bool((grid[index].value >> 7) & uint16_t(1));
 	destGrid[index] = grid[index];
 
 	if (grid[index].value > VOXEL_FREE && isBoundary && noise[index % noiseBufferSize] < erosionProbability)
