+++ MeshFragments/Assets/Shaders/Compute/Fracturer/voxel.glsl
@@ -1,5 +1,5 @@
-#define VOXEL_EMPTY uint8_t(0)		// 0 when a voxel is empy
-#define VOXEL_FREE  uint8_t(1)		// 1 when a voxel is not empty and not occupied by anyone
+#define VOXEL_EMPTY uint16_t(0)		// 0 when a voxel is empy
+#define VOXEL_FREE  uint16_t(1)		// 1 when a voxel is not empty and not occupied by anyone
 
 uniform uvec3	gridDims;
