+++ MeshFragments/Assets/Shaders/Compute/Fracturer/voxelStructs.glsl
@@ -1,4 +1,4 @@
 struct CellGrid
 {
-	uint8_t	value;
+	uint16_t	value;
 };
